fixes to SAM parser, a few typos
git-svn-id: http://biofs04/svn/bioinf/anfo/trunk@2101 93d9a472-daf7-0310-8f21-d7befa554562

+++ src/anfo-sge.cc
@@ -151,7 +151,7 @@ WRAPPED_MAIN
 	poptFreeContext( pc ) ;
 
 	// XXX add adapter trimmer here!
-	for( int i = 0 ; i != more_opts.size() ; ++i )
+	for( size_t i = 0 ; i != more_opts.size() ; ++i )
     {
         if( more_opts[i].first == opt_genome )
         {
